requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
dataclasses-json==0.5.14
decorator==5.1.1
exceptiongroup==1.1.3
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.32
greenlet==2.0.2
grpcio==1.57.0
grpcio-tools==1.57.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
JPype1==1.4.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
konlpy==0.6.0
langchain==0.0.257
langsmith==0.0.19
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.5
numpy==1.25.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
portalocker==2.7.0
protobuf==4.23.4
pyarrow==12.0.1
pydantic==1.10.12
pydeck==0.8.0
Pygments==2.16.1
Pympler==1.0.1
pypdf==3.15.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
qdrant-client==1.4.0
rank-bm25==0.2.2
referencing==0.30.2
regex==2023.6.3
requests==2.31.0
rich==13.5.2
rpds-py==0.9.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==2.0.19
streamlit==1.25.0
tenacity==8.2.2
tiktoken==0.4.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.16
validators==0.20.0
yarl==1.9.2
zipp==3.16.2
